Flutter 2.10.0

+++ lib/providers/userDataProvider.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+
 import '../models/models.dart';
 import '../helpers/helpers.dart';
+import '../constants/constants.dart';
 
 class UserDataProvider with ChangeNotifier {
   String? _accessToken;
@@ -9,11 +11,14 @@ class UserDataProvider with ChangeNotifier {
   bool? _firstSignIn;
   UserProfileAPIBody? _userProfile;
 
+  /// login using the accessToken from SharedPrefences
   Future<bool> accessTokenLogin() async {
     final prefs = await SharedPreferences.getInstance();
     if (!prefs.containsKey('accessToken')) return false;
     var localToken = prefs.getString('accessToken')!;
-    DIOResponseBody response = await API().accessTokenLogin(localToken);
+    DIOResponseBody response = Constants.amplifyEnabled
+        ? await AmplifyAuth.accessTokenLogin(localToken)
+        : await API().accessTokenLogin(localToken);
     if (response.success) {
       _userLoggedIn = true;
       _accessToken = localToken;
@@ -26,28 +31,38 @@ class UserDataProvider with ChangeNotifier {
     }
   }
 
+  /// return the accessToken [accessToken]
   String? get accessToken {
     if (_accessToken == null) return "";
     return _accessToken;
   }
 
+  void refresh() {
+    notifyListeners();
+  }
+
+  /// returns current [loginStatus]
   bool get loginStatus => _userLoggedIn;
 
+  /// return [firstSignIn] status
   bool? get firstSignIn => _firstSignIn;
 
+  /// returns current [userProfile]
   UserProfileAPIBody? get userProfile => _userProfile;
 
-  void assignAccessToken(String token) async {
+  /// update current accessToken and store in SharePrefs with [token]
+  Future<void> assignAccessToken(String token) async {
     if (token != "") {
       _accessToken = token;
       _userLoggedIn = true;
       final prefs = await SharedPreferences.getInstance();
-      prefs.setString('accessToken', token);
-      prefs.setBool('loginStatus', true);
+      await prefs.setBool('loginStatus', true);
+      await prefs.setString('accessToken', token);
       notifyListeners();
     }
   }
 
+  /// update first login status
   void changeFirstLoginStatus(bool status) async {
     _firstSignIn = status;
     final prefs = await SharedPreferences.getInstance();
@@ -55,19 +70,32 @@ class UserDataProvider with ChangeNotifier {
     notifyListeners();
   }
 
+  /// log user out
   void logout(BuildContext context) async {
-    bool response = await API().logout(_accessToken!);
-    if (response || accessToken == null) {
+    if (Constants.amplifyEnabled) {
+      await AmplifyAuth.logout();
       _userLoggedIn = false;
-      notifyListeners();
       _accessToken = null;
       final prefs = await SharedPreferences.getInstance();
-      prefs.clear();
-      Navigator.pushNamedAndRemoveUntil(
-          context, '/welcome', (route) => route.isFirst);
+      if (await prefs.clear()) {
+        Navigator.popUntil(context, (route) => route.isFirst);
+      }
+    } else {
+      var _accessToken = await SharedPrefHelper.accessToken;
+      bool response = await API().logout(_accessToken!);
+      if (response || accessToken == null) {
+        _userLoggedIn = false;
+        _accessToken = null;
+        notifyListeners();
+        final prefs = await SharedPreferences.getInstance();
+        prefs.clear();
+        Navigator.pushNamedAndRemoveUntil(
+            context, '/welcome', (route) => route.isFirst);
+      }
     }
   }
 
+  /// update login status with [status]
   void changeLoginStatus(bool status) {
     _userLoggedIn = status;
     notifyListeners();
